GetActivityForResult and handle the return of the activity.
Refactored the showing card code into a method so it can be reused.
Changed the startActivity for rating to a startActivityForResult.
Coded first stab at going to next card and getting next card set from db when we've done them all.

The going to next card and getting next card is pretty rudimentary right now and might need to be tweaked again.